Safari-609.1.11.1/LayoutTests/media/media-source/media-source-init-segment-duration-expected.txt

This tests the ability of the SourceBuffer to detect the movie fragment duration from a 'mvex' header.

RUN(video.src = URL.createObjectURL(source))
EVENT(sourceopen)
RUN(sourceBuffer = source.addSourceBuffer(loader.type()))
RUN(sourceBuffer.appendBuffer(loader.initSegment()))
EVENT(update)
EXPECTED (video.duration == '10') OK
END OF TEST
